02_REFS_PROTOTYPES/uXPRT/uxprt/pyproject.toml
[project]
name = "uxprt"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Antonio Reid", email = "read.musik@gmail.com"},
]
dependencies = ["uvicorn[standard]>=0.34.2", "python-dotenv>=1.1.0", "fastapi>=0.115.12", "supabase>=2.15.1", "bcrypt>=4.3.0", "h2>=4.2.0", "pydantic[email]>=2.11.4"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
